+++ service/auth.go
@@ -5,6 +5,12 @@ import (
 	"painteer/repository/auth"
 )
 
+type AuthService interface {
+	RegisterUser(user model.CreateUser) (*model.User, error)
+	AuthenticateUser(authId model.AuthId) (*model.User, error)
+	GetUserByID(userId model.UserId) (*model.User, error)
+}
+
 type AuthServiceImpl struct {
 	repo auth.UsersRepository
 }
